Rigify: Clean-up of the IK/FK snapping code.  Should be much more maintainable now.
Also changed rig id generation.  Rig id's now consist of a random
alphanumeric string 8 characters long, with the smallest 8 digits of seconds
since the epoc (in hex) at the time of rig generation appended on the end.
This results in a 16-character string that is ludicrously unlikely to
have any collisions between rigs.  36^8 * 16^8, with the 16^8 being very well
distributed over time.  Ah... paranoia.