Importer ======== The length of the 'unit stick' of the dupliface structure that is used for representing the sticks can now be explicitly chosen before loading a PDB structure.
It was always on 0.05 Angstrom, which is a quite small value. Loading 
structures with lots of sticks was therefore very slow. The value can now be 
changed [from 0 to infinity]. I have set the default value onto 0.2 Angstrom,
which should be sufficient.

Note: The smaller the value, the better is the length precision. However, the 
more single units need to be drawn (more vertices).

By time, I will change the wiki ... in some days.

Blendphys