Rigify bugfix: deleted WGT objects now get properly replaced on generation.
WGT objects were sticking around even when the user deleted them
since they were being used by the bones for visual appearance.  This
meant that if the user deleted the WGT objects in order for Rigify
to regenerate them the next time around, Rigify would fail.

Rigify now checks for this circumstance and handles it appropriately.