Fix own error in rBAc307a89e281a2e8 - OBJ importer *do* have to tessellate/untessellate sometimes.
Issue comes with ngons having holes (i.e. ngons using the same edge more than once).
This is not supported in Blender, so we have to 'work around' it.

However, we can limit this behavior to ngons actually 'blender-invalid', no need to
do that systematically.

So this commit still allows for much better performances with 'usual' ngons.
Those with holes get back to previous performances, more or less.

Many thanks to Campbell for the headup!