Fix T44106: Obj importer broken.
In fact, it's the OBJ file which is utterly broken (thousands of faces reusing the same vertex...).
I think 'breakage' comes actually from change in bmesh (raising error when trying to get
an edge with same vert in both items of the key), afaiks this is not different from
old code in current importer.

Anyway, working around this is easy and cheap, so...

To be backported in final 2.74.