FBX Export: WIP attempt to have 'skip_export' objects.
This would allo us e.g. to not export armature object as a root for bone chains,
or to have a real 'only deform' export option, in each case by ignoring some
nodes and 'reparenting' their children to first exported parent.

Idea is simple, but as always when touching to FBX transform (and even worse,
FBX bones binding), it turns into nightmare (a bit like the 'apply transform' issue).

So stuffing this here for now, in case someone (or myself) feels like
wasting another day or two on this mess - with no guarantee of success...