BVH import: added option to update scene from the BVH.
This commit adds two options to the BVH importer to adjust the scene's
frame rate and duration to that of the BVH file. Since different BVHs
have different frame rates, this makes it possible to import BVH files
for inspection in Blender without having to manually open the BVH file
to find its frame rate and duration. The scene is only extended to fit
the BVH file, and never shortened.

There already exists an option for the opposite, to scale the BVH
animation data to the scene's frame rate. This did not take into account
the scene.render.fps_base property, which is also fixed by this commit.

This closes task T34919.

Reviewers: campbellbarton

Reviewed By: campbellbarton

Maniphest Tasks: T34919

Differential Revision: https://developer.blender.org/D1608