Fix T47255: inset polygon addon left old edges
Not sure why this used to work. The old code used the bmesh faceseq
delete, which kills only the faces. This fix uses the utility
bmesh.ops.delete() with an arg that will maybe kill edges and verts too.
Also made code select all the new polys. This is not ideal (ideal would
be to select only the inner ones), but making the ideal change requires
more work.