Rigify: add new meta rig type pitchipoy.super_palm
D2114 by @icappiello

Scaling non uniformly the hand control on the rig generated with this type will lead to crazy deformations on the finger bones.
The problem is related to some “inherit scale” options on the parented bones.
In order to leave untouched the original Rigify code we created a new rig type and named it “super_palm” using the pitchipoy naming convection.
We changed the bone parenting so that the hand can be scaled non uniformly without destroying its children bones.
This new palm type is needed to append fingers and toes to the “paw” rig type.