Image as planes: remove 'compat' operator, rename 'new' one to standard name.
Add-ons are no API, and are not expected to be used outside of
themselves (i.e. add-ons whall not depend on other add-ons). So no
reason to keep compatibility with previous behavior of import image a
splanes operator (especially since name had to be changed anyway).

Also simplified register/unregister code, exceptions are only to be used
to detect and handle errors, not to handle some valid cases.