Bone Selection Sets: added popup menu for quickly selecting selection set
Pressing Shift+Alt+W pops up a menu with the different bone selection
sets for the current rig. Selecting a set will select all the bones in
that set.

This functionality was copied from [1]; that add-on has already been
used for quite a while in the Blender Animation Studio with great
succes. The main difference is that now the menu uses the selection set
index rather than the name, which means it'll work correctly even when
there are duplicate names.

[1] https://github.com/sybrenstuvel/random-blender-addons/blob/master/selection_set_selector.py