Multi-Objects: UV_OT_export_layout (80%)
"""
Apart from the 'export as PNG' all seems to be working ok.

'Export to PNG' also seems to work but doesn't look acceptable because the mesh
wireframe modifier is used  to draw UV wireframes (I was going to see if a curve
object would be usable to draw wireframes). Something in setting the renderer up
could also be wrong or missing.
"""

D3507 by @Al

Note from reviewer: I can't get this addon to show in the menus. But the syntax
seems correct, and it's pep8 friendly, so no harm in committing it so I can
ask other devs to see why the addon is never registering here.

Also, some changes I made include making it pep8 friendly and replace:
['foo', 'bar'] with {'foo', 'bar'}.