SVG: Properly handle values in exponential notation
Some SVG exporters outputs small values in an exponential
notation. There is no big reason to reject those files.

This change makes it so any notation of the value is accepted.
Only do it in the path point parsing, since other areas are
already dealing with this correct.

Also covered the array parsing covered with a unit test which
can be run as a stand-alone application.

The parsing code is from Jacques Lucke, thanks!

Differential Revision: https://developer.blender.org/D4234