Concept change for dupliverts (atomic structures): the way of naming a duplivert (mesh and obj) has been changed.
Problem: because of changes that were made with respect to instancing
(see https://wiki.blender.org/wiki/Reference/Release_Notes/2.80/Python_API/Scene_and_Object_API)
the mesh and object appear on the 'same level' in the outliner. Finding the mesh and object
was not easy because of the old 'name convention' in 2.79. Now, the names of mesh and object both
start with the element name (e.g., hydrogen) + an identifier of their nature ("_mesh" and "ball"). The
then appear one after the other in the outliner ... .