Fix: the 'CONECT' identifier has been corrected.
On October 19th 2018, the 'CONECT' identifier was changed to 'CONNECT'. This is wrong
because in a PDB file a stick is identified as 'CONECT'. I recognized that this morning
and could not load any sticks. I have put a note into the code. Please leave it as is!
There are still some errors afterwards, which I will solve later on.