Concept change: massive use of the collection principle to well structure atoms and sticks of the molecule
Thanks to the great concept of 'collections', which Blender 2.80 is proposing, all parts of
the molecule (atoms and sticks) can be organized in an hierarchical order in dependence
on the chemical elements. Atoms (balls) are further separated from sticks (cylinders). For
some molecules, I tested all possible properties one can chose during the import, and it
works very well.