Fix T67266: .obj import error with mtllib name in quotes - regression from 2.79b.
While this is not officialy part of OBJ format, some softwares 'solve'
the spaces-in-filenames issue by delimiting those with quotes.

Since this is not too much a hassle, let's add back support for this.

Based on work from Robert Guetzkow (@rjg), thanks.

Note: not sure whether this should be ported to 2.80 release, while this
change should be reasonably safe, it's not an obvious one-liner either,
and the issue is not really critical (don’t think any major software
uses that trick?)...