animation_animall: Fix T69509 Quick smoke broken
The previous commit rBA0a6d3d725f0b broke quick smoke by adding a
handler to force viewport refresh. This commit adds a check
to refresh only data which have animation data. So quick smoke
is not really fixed but the bug happens only if the data is animated.