SVG Import: Fix offset while import SVG files
The file saved as Inkscape SVG moved to the top, regular SVG stays in place. Blender checks if SVG have an special attribute  `inkscape:version`, and if it has one, it move all SVG to the top. The idea is to match bottom right corner with world origin, instead top right corner:

But why height is not equal the real height of the SVG?  Well, because it's not a height of SVG itself, it is size of SVG on printing or displaying in web-page or in previewer. The real height of SVG in SVG-units is located in viewbox attribute:

https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/viewBox

So I suggest to move SVG content to real height of SVG located in viewbox attribute:

Reviewers: sergey, antoniov

Reviewed By: antoniov

Subscribers: meta-androcto, antoniov, jms, sergey

Tags: #add-ons

Differential Revision: https://developer.blender.org/D5727