Rigify: make the generic FK to IK snap operator even more generic.
The operator itself simply snaps a chain of bones to a different
chain, so nothing in it in fact is specific to IK or FK. To make
this even more explicit, rename the operator and add some extra
options to control the tooltip and which properties are changed.

Also include some other minor enhancements in the script utilities.