Pie Menus: Update sculpt icons.
Update the icons of the sculpt pie menu to be consistent with the new
sculpt icons.

The icons currently don't respect UI sizes, so we have to increase the
menu sizes for them to fit. We also add white spaces at the start of
the text to avoid icon-text intersections.

Reviewed By: Brendon Murphy

Differential Revision: https://developer.blender.org/D6373