PDT: Modify Fillet functionality
This enables two non-connected edges to be filleted. The Process starts with an
intersection of the two edges, then fillets the corner according to input values
in the UI. This option is selected by setting the Int/Fillet checkbox.

The command Line version now takes the `fi...` format to intersect first. e.g.

`fi1.1,6,0.05` performs an intersected fillet of 1.6 units radius, 6 segments,
concave profile.