PDT: Move stray files from add-on root folder
The error likely came about because we first commit files to @clockmender's
upstream github repo and then use git format-patch to create patches which
are then applied to the blender add-on repo with:

`git-am -3 --rerere-autoupdate foo.patch`.

Here, the issue seems to be that we've renamed files in the upstream github
repo, but that I then failed to first create and commit empty files with
the same name in the precision_drawing_tools/ repo, which meant that git
did the best it could and added them to the root of the blender add-ons
repo since the files didn't exist in the add-on repo index when they were
added.