Initial VR Pose bookmarking system
Idea is to have a list of VR pose bookmarks in the sidebar, so you can
jump back and forth between custom, pinned poses (rotation+location).
Right now the bookmarks have to be a camera, whereby you can choose
between following the active scene camera, or a custom fixed one.
Adding completely custom locations (not bound to a camera or any other
object) will also be added, but needs changes in RNA first.

Note that the latest state of this is mostly untested. I'm travelling
and can't test with an actual VR session.