Collection Manager: Add QCD System. Task: T69577
Adds a Quick Content Display (QCD) system to the Collection Manager.
This consists of a 3D View Header widget and a floating panel
similar to the layers system in blender 2.7x, along with
hotkeys to view/move objects to QCD slots, and additions to the
main Collection Manager popup to allow you to manage which collections
correspond to which slots.