OBJ: Fix mistakes in rBAc03185fe40dc,  add some tooltips.
Please do not commit random things to code maintained by other people,
especially if you are breaking basic 101 things like codestyle. This is
loss of time for everybody. Even worse since not ansewring to comments
on original commit.

Also no need to add extra `text` parameters (and more useless UI
messages) in add-ons, when the only usage of the property's name itself
is in own add-on' UI...