PDT: Fix Library path issue, Add Trig Waves Functions
An issue occurred if you opened Blender with a new file where the
Parts Library path pointed relative to the Blender App, not the working
directory, so was not found.

The Parts Library file is now moved to the Parts Library Menu, so must
be set for new .blend files. Existing .blend files are not affected by
this change.

Error message altered to reflect the option of having this Parts Library
live anywhere.

New Feature: Add Trig Waves Functions

Requires the user to select an existing object in the Trig menu.