Collection Manager: Fix scrolling issue. Task: T69577
Set the selection to the collection when clicking on its expander.
This prevents the tree view scrolling changing when expanding/collapsing
subcollections.

This also sets the selection to the first collection when collapsing
all collections to prevent loosing the selected row and then later
adding collections to some unknown region.