Fix T74483: Rigify Meta-Human Neck Bone stretching.
This implements Ivan's suggestion: all the constraints on
spine bones have to be moved from ORG to its DEF bone. All
the spine ORG-bones have to be unconnected-children of their
corresponding tweak.

This is actually how chain rigs work in my own Rigify feature set,
so as far as I'm concerned this is a tried and viable solution.
The result is that, if you want to parent something to your chain
rig, you can do it in two ways: If you parent it to the ORG bone
(ie. by default), it will not inherit any squash and stretch induced
by the "tweak" bones. If you parent it to the DEF bone however,
it will behave the same as before.

This raises a backwards compatibility concern; If you want the old
behavior, you have to express that in your metarig explicitly, by
parenting your bone to the DEF bone in some way. This patch also
only affects the spine rig, which makes it inconsistent with other
chain rigs in Rigify in this regard.

Maniphest Tasks: T74483

Differential Revision: https://developer.blender.org/D7801