UI: Dont abbriviate location & rotation
This resolves one of the last few areas where we still use inappropriate 
abbreviations. Reading abbreviated words is usually slower, because 
users must parse, guess and translate the words. Using abbreviations 
such as 'rot' is also especially bad since it's a word in itself too.

The main advantage of abbreviations is that they are faster to *write*, 
which just isn't a concern for text in the UI.