Collection Manager: Object selection. Task: T69577
Adds the ability to select all object in a QCD slot.
Alt+LMB deselects everything and selects all objects in the QCD slot.
Alt+Shift+LMB adds/removes objects in the QCD slot to/from the selection.

Added a selection operator for use in the Collection Manager popup.
(currently unused)