VR Scene Inspection: Better way to calculate landmark angle from viewer
When using the "Add VR Landmark from Session" operator, we want to
calculate an angle around the Z-axis for the landmark to use as
ground-aligned base rotation. The former way to calculate this was
rather primitive and broke easily (angle would point in the opposite
direction).
Now we project the camera direction onto the ground plane to get the
angle. This should be a very reliable way to do it.