Rigify: Code Cleanup: Use ControlLayersOption class
This class exists specifically for the purpose of defining and drawing the UI for bone layer assignment parameters, as seen in other rig types like super_chain, limb_rigs, simple_tentacle, etc. Only super_face was missing it for some reason.

Although the way it is used may be questionable, it's consistent with the rest of the codebase.

No functional changes.

Note for future: I would also like to use the ControlLayersOption class in even more places, and also maybe rename it and make improvements to it as needed. This would also be part of a bigger design of how Rigify handles bone organization, which I should write down.

Reviewed By: sybren

Differential Revision: https://developer.blender.org/D8802