Bone Selection Sets: USE_INSERTION on CollectionProperties
Use the new 'USE_INSERTION' override flag (rBdb314ee7a472) to allow creating new entries of Selection Sets on overridden rigs. I tested including saving and reloading, assigning and removing bones from a set:

- Can create and name new, local sets
- Can add and remove bones to and from those local sets
- Can add bones to original, non-local sets and remove those same bones
- Can NOT remove original sets
- Can NOT remove originally assigned bones bones from originally existing sets

To me this all seems to be working as intended, so might as well use it!

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D8971