Collection Manager: Add disable override. Task: T69577.
Add an override checkbox to the preferences to override
the object hiding hotkeys with object disabling hotkeys
and add menu items to the Object->Show/Hide menu for the
disable object operators.

Disabling objects is needed to preserve their visibility
state when excluding and unexcluding collections.
This makes it easier to disable objects, and disabling
objects prevents QCD slot switching from resetting the
objects visibility state.