Rigify: Turn Rotation Mode converter into single operator
This operator used to demand a huge chunk for itself as the sole inhabitant of the Rigify panel in the sidebar with a bunch of options that are only relevant when the operator is actually ran. Also for some reason it was split up to two separate operators for the sake of having "convert all actions" vs "convert current action" buttons.

So, I removed that UI from the sidebar entirely, turned all the options into parameters on a SINGLE operator which asks for those options when ran using invoke() and draw(). This operator now lives in the Pose drop-down menu, which is where you would expect to find it. This is also searchable with the new search menu, unlike the sidebar.

Also note that this operator has precisely //nothing// to do with Rigify, and should be a separate addon. I would split it out in a separate patch.

Reviewed By: angavrilov

Differential Revision: https://developer.blender.org/D9096