Rigify: Addon Prefs & Feature Sets UI
This patch removes the two drop-down menus from the Rigify addon preferences, replacing it with a checkbox and a UIList.

Before:
{F8766280}

After:
{F8791382}

There is also a small bit of new functionality. The following keys are now recognized in the rigify_info dictionary, similar to Blender addons:
`'link', 'doc_url', 'tracker_url', 'description', 'author', 'version', 'warning'.`
They will be displayed underneath the UIList when the feature set is the active one in the list.

**Ideas for future**
I think this UI better enables us to do some cool things in the future:
- Disable feature sets without uninstalling them.
- Have a base set of feature sets in the list that aren't actually installed. Instead, selecting them would just reveal a button that opens their repository URL. This could be used for officially supported but external feature sets, such as what Legacy should be.

Feedback welcome.

Reviewed By: angavrilov

Differential Revision: https://developer.blender.org/D8519