FBX export: Add an option to disable global space transform.
Allow to disable the rotation matrix that is applied to the object transforms
during export.

With this option disabled only the axis system is written to the fbx file,
but the object transforms are left as-is. This leaves it up to the importer
on the other side to apply the space transform during import.

Unity has added an import option to apply space transform on import in its
latest version, but the current version of setting the axis system in the
fbx file and applying the matrix causes unexpected behaviour.

Most users will expect that Blender has a forward direction of -Y and
an up direction of +Z (Suzanne is looking in the -Y direction and the
"front" view is coming from the -Y direction). But if you set the axis
conversion to -Y and +Z the exporter will apply a 180° rotation because it
assumes a forward direction of +Y.

When done this way, using the new Unity import setting, the mesh will be
imported correctly in Unity, but the rotation of the root objects will
contain that 180° rotation. Using the +Y and +Z axes during export will
import everything without any rotations, but what the user expects to be
forward will now be pointing the other direction in Unity.

When we just write the axis system as -Y and +Z to the fbx file and leave
the rotations of the root objects untouched, we can effectively define any
foward axis and have it still be imported correctly, because every conversion
will be done on Unitys side.

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D8078