Fix T83099: Greasepencil add-on keymap item removal
The greasepencil add-on removed keymap items it hadn't created
itself, thus invalidating references that other add-ons held for
the keymap items they've created. This is fixed by storing a
reference to the created keymap items alongside the keymap and
only removing those items that the add-on created itself.
Additionally, this commit fixes the doc string that was no
longer valid.

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D9670