Fix T83386: Use type annotation instead of assignment
The Sapling Tree Gen add-on previously assigned the `FloatProperty`
to `bend`. This caused accessing `bend` in calculations to return
a reference to the `FloatProperty` instead of its value. This
commit replaces the assignment with an annotation as it is the
correct usage since the Python API changes for Blender 2.80.

NOTE: The add-on currently has another problem T83360 which
prevents it from working. For the implementation of the patch
I had to comment out the usages of `props.leafDupliObj`.

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D9747