Rigify: use separate widget collection for each generated rig.
As pointed out in comments to T73114, using the same name has downsides.

In addition, remove the widget parent object that was inherited
from the old pre-collection way Rigify worked and has no other
purpose than grouping objects like the collection already does.

Rename the widgets and the collection when renaming the rig.

Finally as an aside add a couple of options to create_widget.