BlenderKit: upload displayName
This enables 2 names for assets - displayName is shown to user,
but name is the one saved in the assets. 90% of time these will be the same, but when a user only updates the name of the asset,
there is no more need to reupload because the name needed to be the same in the .blend file.
also rename draw_callback_2d_search to draw_asset_bar