Fix T84616: Power Sequencer 2.90 API update
The Power Sequencer wasn't updated to use the 2.90 API which
changed the selection operators for the sequencer. This commit
replaces the old operator with the new
`bpy.ops.sequencer.select_side_of_frame` to achieve the same
functionality as before.

Credit for the identification and fix suggestion goes to @kreako

Reviewed By: campbellbarton

Differential Revision: https://developer.blender.org/D10089