BlenderKit: fix colorspace with filmic addon
This currently keeps colorsettings for images to default, because we can hardly predict the names of possible colorspace settings added in variants of this or other addons.
All colorsettings settings were moved to a set_colorspace function in image_utils.py and everything happens in a try statement.
Scenes - fix link/append support and add it to UI, further support switching to scene after finishing the download.
Fix drag-drop of materials to linked objects - now theres an explanation pop-up.