Fix SVG material cache bug
This fixes a bug that causes a new material to be created for every color even if it has been cached. The color name is transformed and then used as a key for the cache, but the cache was being checked against the pre-transformed version, causing it to fail every time. This fix makes sure the key being looked for in the cache has been transformed, and can therefore be found as expected.