Pose Library: add operator to easily go back to previous Action
Creating a pose Action switches the object's active Action to the newly
created one (for inspection/tweaking). This now also enables a button that
can be pressed to go to the previous Action. This makes it much easier to
line up various poses in an action and create pose assets from them.