Attempt to address T67111.
If you have a relative path (e.g. //../objects/object.blend) and you use
bpy.path.abspath, it leaves the relative part (the double dots) as part
of the path (e.g. ~/some_project/scenes/../objects/object.blend). While
Windows is perfectly happy to open files with this kind of path, it
seems to get confused about paths that are relative to that. The fix is
to wrap the bpy.path.abspath function with os.path.abspath and
everything appears to remain sane.

Also introduced a small fix for the new linked node groups feature; in
2.93, the Item tab in the sidebar has been renamed to Node.