Rigify: Make operators searchable by adding a Rigify header drop-down
This or something like this is necessary because the new search menu doesn't search the sidebar nor the properties panel for UI buttons. I think the ability to run Add Sample and Generate from the search menu is useful, so being able to do it without having Developer Extras enabled is important.

I would also like to propose some new operators in the future, and this patch sets up a nice spot for them to go into.

I didn't remove any buttons in this patch, but would like to do in future, so these things will once again only exist in one place (which would be the header).

Reviewed By: angavrilov

Differential Revision: https://developer.blender.org/D9094