Collection Manager: CM popup UX fixes. Task: T69577
Renamed the Set Collection Operator to Send Objects To Collection
and updated it's tooltip.
Changed it's icon for a collection with no objects
from the grid icon to the import icon for (hopefully)
better discoverability.

Updated the filter bar's invert icon to use the standard arrows icon.
Changed the filter collections by selected objects icon to the selected
objects icon used by the send objects to collection operator.