Pose library: show message when "Copy Pose As Asset" is pressed
Show a message in the UI when "Copy Pose As Asset" was successful, to
explain where the corresponding "paste" button can be found.

Since the "copy" is asset-type specific, this one is located in the
animation editor. However, the "paste" button is generic for any asset
type, and thus it is located in the asset browser itself.