New add-on: Pose Library v2.0
Add the Pose Library 2.0 add-on, which is based on the Asset Browser. It
heavily depends on functionality of Blender itself, most notably in
`pose_lib_2.c`.

This add-on will be enabled by default by Blender, in order to be a proper
replacement for the old, build-in pose library. It can also convert pose
libraries from the old to the new one.