Pose Library: change how cleanup is done after Copy As Asset
Copy As Asset creates an asset datablock, saves it to disk, and then
removes it again. This removal has a check to ensure the temp datablock
isn't accidentally still in use by something. When this check fails, it now
still forces the cleanup. The message is now there just to ask people to
file a bug report, instead of blocking their workflow altogether.