Rigify: unhide the rig before generation if it is hidden.
Rig generation fails if the target rig that is to be overwritten
is hidden, either directly or via collections. To improve usability,
try unhiding the object, and fail with an appropriate message if
it doesn't work.

In addition, slightly refactor code that selects a suitable
unhidden collection to use for generation.

This started as D11355 by @Mets, but was somewhat refactored.