Collection Manager: Improve UI for adding QCD slots. Task: T69577
Replace the X labels in the QCD header widget with operators that
generate the corresponding slot when clicked, allows for moving
objects to the new slot by pressing hotkeys when clicking.  Each
unassigned slot operator has a tooltip specifying it's an
unassigned slot, which slot number it corresponds to, and lists
the hotkeys that can be used with it and their functions.

Adds a menu item to the QVT menu (the eye icon beside the QCD
header widget) to add any missing QCD slots so that you end up
with a full 20 slots.