MathVis: Mark the selected display item with an active color
When the number of items defined in MathVis grows, then finding
a specific item in the ViewPort becomes more difficult. One way
to "solve" this is by limiting the number of displayed items.

This patch adds another way by color marking the selected item
in the MathVis list widget with an 'active' color
I have implemented this for matrix bounding boxes and vectors.

Benefit: this way you can quickly step over the list and see
which item highlights in the viewport.

Example:
{F10357951}

Reviewed By: campbellbarton

Differential Revision: https://developer.blender.org/D12360