T91111: possible out of range when iterating over Math Vis item list
While preparing T91111 i found an issue with out of date index after my recent changes.
The problem is:

bpy.context.window_manager.MathVisProp can be "0" even when the list is actually empty.

This patch fixes the index issue, although the true cause might still be found elsewhere.
there is more to fix, so this is only a first step to a complete fix.

Reviewed By: campbellbarton

Differential Revision: https://developer.blender.org/D12390