Pose Library: avoid breakage due to changes in the asset browser
Asset browser moved from Categories (the big, fixed buttons) to Catalogs
(the tree-based, user-creatable organisation system). This required some
changes to the Pose Library code, as that was still expecting to be able
to use the categories system.

The entire workflow has to be looked at better, but at least now it's not
broken anymore.