Pose Library: only allow pasting assets when in Current File library
Only allow the Paste Asset button to work when the asset browser is set
to the "Current File" library. This ensures that the pasted asset can
actually be seen.

Without this change, an artist is likely to assume the asset will be
pasted into the shown library. This is not the case, though -- pasted
assets always go into the current blend file.